pyproject.toml
[project]
name = "gold-stock-checker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "curl-cffi>=0.14.0",
    "python-dotenv>=1.2.1",
    "selectolax>=0.4.6",
    "structlog>=25.5.0",
    "tenacity>=9.1.2",
]
